Explicitly cast to an uint8_t the result

+++ include/parser.hpp
@@ -24,7 +24,8 @@ namespace chip_8
 
     std::array<uint8_t, 2> constexpr separate_byte(uint8_t byte) noexcept
     {
-        return { byte >> 4, byte & 0x0F };
+
+        return { static_cast<uint8_t>(byte >> 4), static_cast<uint8_t>(byte & 0x0F) };
     }
 
     auto constexpr get_nibbles(std::ranges::viewable_range auto&& bytes) noexcept
